refactor: onAttack method, some change.
